# apifox/floworder.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FlowOrder API",
    "description": "FlowOrder 高并发预约交易与履约一致性平台接口集合。主业务接口优先走 gateway 8088；internal/recovery 和 internal/mq 接口用于本地恢复验证，默认需要 floworder.admin.enabled=true。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8088",
      "description": "Gateway - 主业务入口"
    },
    {
      "url": "http://127.0.0.1:8081",
      "description": "resource-service 直连 - recovery/resource admin"
    },
    {
      "url": "http://127.0.0.1:8082",
      "description": "order-service 直连 - order admin"
    },
    {
      "url": "http://127.0.0.1:8720",
      "description": "Sentinel transport 示例端口"
    }
  ],
...
  "paths": {
...
    "/internal/recovery/cases/inspect": {
      "get": {
        "summary": "OrderCare 案例事实诊断",
        "description": "按 requestId/orderNo/deductNo/deadLetterId 聚合权威事实。需要 floworder.admin.enabled=true。",
        "tags": ["恢复控制面"],
        "parameters": [
          {
            "name": "identifierType",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "enum": ["REQUEST_ID", "ORDER_NO", "DEDUCT_NO", "DEAD_LETTER_ID"]},
            "example": "REQUEST_ID"
          },
          {
            "name": "identifierValue",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "example": "ORDERCARE-M05-REQUEST"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {"application/json": {"schema": {"type": "object", "properties": {
              "code": {"type": "integer", "example": 200},
              "message": {"type": "string", "example": "success"},
              "data": {"$ref": "#/components/schemas/RecoveryCaseResult"}
            }}}}
          }
        }
      }
    },
    "/internal/recovery/proposals": {
      "post": {
        "summary": "创建不可变恢复 Proposal",
        "description": "使用稳定 proposalId 幂等创建预演，不修改订单或库存终态。",
        "tags": ["恢复控制面"],
        "responses": {
          "200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {
            "code": {"type": "integer", "example": 200},
            "message": {"type": "string", "example": "success"},
            "data": {"$ref": "#/components/schemas/RecoveryProposalResult"}
          }}}}}
        },
        "requestBody": {"required": true, "content": {"application/json": {
          "schema": {"$ref": "#/components/schemas/RecoveryProposalCreateRequest"},
          "example": {"proposalId": "prop-apifox-001", "identifierType": "REQUEST_ID", "identifierValue": "ORDERCARE-M05-REQUEST", "actionType": "REPLAY", "suggestedReason": "ORDER_TIMEOUT 库存释放消息进入死信"}
        }}}
      }
    },
    "/internal/recovery/proposals/{proposalId}": {
      "get": {
        "summary": "查询恢复 Proposal",
        "description": "返回 Proposal、Action 命令状态和独立的业务结果状态。",
        "tags": ["恢复控制面"],
        "parameters": [{"name": "proposalId", "in": "path", "required": true, "schema": {"type": "string"}, "example": "prop-apifox-001"}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {
          "code": {"type": "integer", "example": 200},
          "message": {"type": "string", "example": "success"},
          "data": {"$ref": "#/components/schemas/RecoveryProposalResult"}
        }}}}}}
      }
    },
    "/internal/recovery/proposals/{proposalId}/execute": {
      "post": {
        "summary": "执行已审批的恢复 Proposal",
        "description": "审批必须绑定当前 Proposal 版本、状态指纹和摘要；executionOwner 用于 Action 租约。",
        "tags": ["恢复控制面"],
        "parameters": [{"name": "proposalId", "in": "path", "required": true, "schema": {"type": "string"}, "example": "prop-apifox-001"}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {
          "code": {"type": "integer", "example": 200},
          "message": {"type": "string", "example": "success"},
          "data": {"$ref": "#/components/schemas/RecoveryProposalResult"}
        }}}}}},
        "requestBody": {"required": true, "content": {"application/json": {
          "schema": {"$ref": "#/components/schemas/RecoveryProposalExecuteRequest"},
          "example": {"proposalVersion": 1, "stateFingerprint": "从 Proposal 复制", "effectsDigest": "从 Proposal 复制", "warningsDigest": "从 Proposal 复制", "previewDigest": "从 Proposal 复制", "approvalId": "approval-apifox-001", "approvedBy": "operator-li", "approvalComment": "已核对影响和警告", "executionOwner": "tool-execution-apifox-001"}
        }}}
      }
    },
    "/internal/recovery/actions/{actionRequestId}": {
      "get": {
        "summary": "查询 Recovery Action 权威状态",
        "description": "用于 execute 响应丢失或 Runtime 重启后的原 Action 对账。",
        "tags": ["恢复控制面"],
        "parameters": [{"name": "actionRequestId", "in": "path", "required": true, "schema": {"type": "string"}, "example": "act-from-proposal"}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {
          "code": {"type": "integer", "example": 200},
          "message": {"type": "string", "example": "success"},
          "data": {"$ref": "#/components/schemas/RecoveryActionResult"}
        }}}}}}
      }
    },
    "/internal/recovery/actions/{actionRequestId}/reconcile": {
      "post": {
        "summary": "确定性对账或接管 Recovery Action",
        "description": "只使用原 actionRequestId；活跃租约等待，过期租约按业务事实 CAS 接管。",
        "tags": ["恢复控制面"],
        "parameters": [{"name": "actionRequestId", "in": "path", "required": true, "schema": {"type": "string"}, "example": "act-from-proposal"}],
        "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {
          "code": {"type": "integer", "example": 200},
          "message": {"type": "string", "example": "success"},
          "data": {"$ref": "#/components/schemas/RecoveryActionResult"}
        }}}}}},
        "requestBody": {"required": true, "content": {"application/json": {
          "schema": {"$ref": "#/components/schemas/RecoveryActionReconcileRequest"},
          "example": {"executionOwner": "runtime-resume-apifox-001"}
        }}}
      }
    },
    "/internal/recovery/reservation/check": {
      "get": {
        "summary": "V10 预约链路核对",
        "description": "直连 resource-service 8081。需要 floworder.admin.enabled=true。",
        "tags": [
          "恢复控制面"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "v12-confirm-164232"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReservationRecoveryCheckResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/recovery/dead-letter/preview": {
      "post": {
        "summary": "V10 死信恢复预览",
        "description": "只预览并写审计日志，不执行恢复动作。直连 resource-service 8081。",
        "tags": [
          "恢复控制面"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RecoveryPreviewResult"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryDeadLetterRequest"
              },
              "example": {
                "actionRequestId": "v10-preview-001",
                "deadLetterId": 2070417281797263362,
                "actionType": "REPLAY",
                "operator": "admin",
                "reason": "preview before replay",
                "force": false
              }
            }
          }
        }
      }
    },
    "/internal/recovery/dead-letter/execute": {
      "post": {
        "summary": "V10 死信恢复执行",
        "description": "会调用 replay/ignore，需谨慎。建议先 preview。直连 resource-service 8081。",
        "tags": [
          "恢复控制面"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RecoveryExecuteResult"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryDeadLetterRequest"
              },
              "example": {
                "actionRequestId": "v10-execute-001",
                "deadLetterId": 2070417281797263362,
                "actionType": "REPLAY",
                "operator": "admin",
                "reason": "manual replay after check",
                "force": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReservationRequestResult": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "description": "0待处理 10处理中 20成功 30待重试 40失败 50人工确认"
          },
          "orderNo": {
            "type": "string"
          },
          "orderStatus": {
            "type": "integer",
            "description": "10已预约 20已确认 30已取消 40已超时"
          },
          "latestOrderEventType": {
            "type": "string"
          },
          "latestOrderEventTime": {
            "type": "string",
            "format": "date-time"
          },
          "orderEventVersion": {
            "type": "integer"
          },
          "retryCount": {
            "type": "integer"
          },
          "lastError": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderQueryResult": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "orderNo": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
...
      "RecoveryCaseResult": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "schemaVersion": {"type": "string", "example": "floworder-recovery-case-v1"},
          "caseKey": {"type": "string"},
          "identifierType": {"type": "string"},
          "identifierValue": {"type": "string"},
          "diagnosisCode": {"type": "string"},
          "factsComplete": {"type": "boolean"},
          "recoveryEligible": {"type": "boolean"},
          "evidence": {"type": "array", "items": {"type": "string"}},
          "hardRisks": {"type": "array", "items": {"type": "string"}}
        }
      },
      "RecoveryProposalCreateRequest": {
        "type": "object",
        "required": ["proposalId", "identifierType", "identifierValue", "actionType"],
        "properties": {
          "proposalId": {"type": "string"},
          "identifierType": {"type": "string"},
          "identifierValue": {"type": "string"},
          "actionType": {"type": "string", "enum": ["REPLAY"]},
          "suggestedReason": {"type": "string"}
        }
      },
      "RecoveryProposalExecuteRequest": {
        "type": "object",
        "required": ["proposalVersion", "stateFingerprint", "effectsDigest", "warningsDigest", "previewDigest", "approvalId", "approvedBy", "executionOwner"],
        "properties": {
          "proposalVersion": {"type": "integer"},
          "stateFingerprint": {"type": "string"},
          "effectsDigest": {"type": "string"},
          "warningsDigest": {"type": "string"},
          "previewDigest": {"type": "string"},
          "approvalId": {"type": "string"},
          "approvedBy": {"type": "string"},
          "approvalComment": {"type": "string"},
          "executionOwner": {"type": "string"}
        }
      },
      "RecoveryProposalResult": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "schemaVersion": {"type": "string", "example": "floworder-recovery-proposal-v1"},
          "proposalId": {"type": "string"},
          "proposalVersion": {"type": "integer"},
          "proposalStatus": {"type": "string"},
          "actionRequestId": {"type": "string"},
          "actionStatus": {"type": "string"},
          "caseOutcome": {"type": "string"},
          "stateFingerprint": {"type": "string"},
          "previewDigest": {"type": "string"},
          "canExecute": {"type": "boolean"},
          "expiresAt": {"type": "string", "format": "date-time"}
        }
      },
      "RecoveryActionReconcileRequest": {
        "type": "object",
        "required": ["executionOwner"],
        "properties": {"executionOwner": {"type": "string"}}
      },
      "RecoveryActionResult": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "schemaVersion": {"type": "string", "example": "floworder-recovery-action-v1"},
          "proposalId": {"type": "string"},
          "actionRequestId": {"type": "string"},
          "actionStatus": {"type": "string"},
          "caseOutcome": {"type": "string"},
          "reconciliationStatus": {"type": "string"},
          "executionOwner": {"type": "string"},
          "executionLeaseUntil": {"type": "string", "format": "date-time"},
          "leaseExpired": {"type": "boolean"},
          "reconcileCount": {"type": "integer"},
          "reconciledAt": {"type": "string", "format": "date-time"}
        }
      },
      "RecoveryDeadLetterRequest": {
        "type": "object",
        "required": [
          "actionRequestId",
          "deadLetterId",
          "actionType",
          "operator"
        ],
        "properties": {
          "actionRequestId": {
            "type": "string",
            "example": "v10-preview-001"
          },
          "deadLetterId": {
            "type": "integer",
            "format": "int64",
            "example": 2070417281797263362
          },
          "actionType": {
            "type": "string",
            "enum": [
              "REPLAY",
              "IGNORE"
            ],
            "example": "REPLAY"
          },
          "operator": {
            "type": "string",
            "example": "admin"
          },
          "reason": {
            "type": "string",
            "example": "preview before replay"
          },
          "executionOwner": {
            "type": "string",
            "example": "tool-execution-001"
          },
          "force": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "RecoveryPreviewResult": {
        "type": "object",
        "properties": {
          "actionRequestId": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "targetType": {
            "type": "string"
          },
          "targetKey": {
            "type": "string"
          },
          "canExecute": {
            "type": "boolean"
          },
          "currentStatus": {
            "type": "integer"
          },
          "recommendedAction": {
            "type": "string"
          },
          "effects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deadLetter": {
            "$ref": "#/components/schemas/MqDeadLetterAdmin"
          }
        }
      },
      "RecoveryExecuteResult": {
        "type": "object",
        "properties": {
          "actionRequestId": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "targetType": {
            "type": "string"
          },
          "targetKey": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "恢复命令状态；SUBMITTED 只表示命令已提交，不表示业务已收敛",
            "enum": [
              "EXECUTING",
              "SUBMITTED",
              "FAILED",
              "IDEMPOTENT_SUBMITTED"
            ]
          },
          "message": {
            "type": "string"
          },
          "executedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReservationRecoveryCheckResult": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "reservationRequest": {
            "$ref": "#/components/schemas/ReservationRequestResult"
          },
          "deductRecord": {
            "type": "object",
            "additionalProperties": true
          },
          "order": {
            "$ref": "#/components/schemas/OrderQueryResult"
          },
          "orderQueryError": {
            "type": "string"
          },
          "stockItem": {
            "type": "object",
            "additionalProperties": true
          },
          "quota": {
            "type": "object",
            "additionalProperties": true
          },
          "inventoryDiff": {
            "type": "integer"
          },
          "inventoryInvariantOk": {
            "type": "boolean"
          },
          "resourceOutboxes": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "unresolvedDeadLetterCount": {
            "type": "integer",
            "format": "int64"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MqDeadLetterAdmin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "messageId": {
            "type": "string"
          },
          "deadQueue": {
            "type": "string"
          },
          "producerService": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "bizKey": {
            "type": "string"
          },
          "exchangeName": {
            "type": "string"
          },
          "routingKey": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "deathReason": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "description": "0待处理 10重放中 20已解决 30已忽略"
          },
          "replayCount": {
            "type": "integer"
          },
          "lastError": {
            "type": "string"
          },
          "replayedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "handledBy": {
            "type": "string"
          },
          "resolutionNote": {
            "type": "string"
          }
        }
      },
...
    }
  }
}